requirements.txt
Flask==2.1.2
marshmallow==3.16.0
web3==5.29.2
Flask-RESTful==0.3.9
Flask-HTTPAuth==4.7.0
sentry-sdk==0.10.2
pydash==5.0.0
celery==5.2.7
python-dotenv==0.16.0
Flask-PyMongo==2.3.0
redis==3.1.0
gunicorn==20.1.0
pymongo==4.1.1
apscheduler==3.9.1
boto3==1.21.38
motor==3.0.0
redis-py-cluster==2.1.3
